more bug fixes; todo: sampling is at the right frequency and is inconsistent

+++ firmware/HelperClasses/AmpScalingMath.h
@@ -3,4 +3,4 @@
 #include <stdint.h>
 #include <stdbool.h>
 
-uint16_t AmpToBrightness(int32_t sin, int32_t cos, uint8_t numCycles);
+uint8_t AmpToBrightness(int32_t sin, int32_t cos, uint8_t numCycles);
